+++ sdk/internal/recording/matchers.go
@@ -12,13 +12,13 @@ import (
 	"testing"
 )
 
-// Optional parameters for the AddBodilessMatcher operation
+// Optional parameters for the SetBodilessMatcher operation
 type MatcherOptions struct {
 }
 
-// AddBodilessMatcher adjusts the "match" operation to exclude the body when matching a request to a recording's entries.
+// SetBodilessMatcher adjusts the "match" operation to exclude the body when matching a request to a recording's entries.
 // Pass in `nil` for `t` if you want the bodiless matcher to apply everywhere
-func AddBodilessMatcher(t *testing.T, options *MatcherOptions) error {
+func SetBodilessMatcher(t *testing.T, options *MatcherOptions) error {
 	if recordMode != PlaybackMode {
 		return nil
 	}