+++ sdk/internal/CHANGELOG.md
@@ -3,10 +3,10 @@
 ## 0.8.3 (Unreleased)
 
 ### Features Added
-* Adds the `AddBodilessMatcher` method [#16256](https://github.com/Azure/azure-sdk-for-go/pull/16256)
+* Adds the `SetBodilessMatcher` method [#16256](https://github.com/Azure/azure-sdk-for-go/pull/16256)
 ### Breaking Changes
 
-### Bugs Fixed
+### Breaking Changes
 * Renames `ResetSanitizers` to `ResetProxy`
 
 ### Other Changes